Change persistence package
