+++ build.sbt
@@ -225,7 +225,7 @@ def runExampleWithOtelAgent = Command.command("runExampleWithOtelAgent") { state
     Seq(
       run / javaOptions ++= Seq(
         s"-javaagent:$root/opentelemetry-javaagent110.jar",
-//        s"-Dotel.javaagent.debug=true",
+        s"-Dotel.javaagent.debug=true",
         s"-Dotel.service.name=mesmer-example",
         s"-Dotel.metrics.exporter=otlp",
         s"-Dotel.metric.export.interval=5000",
