+++ otel-extension/src/main/java/io/scalac/mesmer/otelextension/MesmerAkkaPersistenceInstrumentationModule.java
@@ -3,7 +3,7 @@ package io.scalac.mesmer.otelextension;
 import com.google.auto.service.AutoService;
 import io.opentelemetry.javaagent.extension.instrumentation.InstrumentationModule;
 import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
-import io.scalac.mesmer.agent.akka.persistence.AkkaPersistenceAgent;
+import io.scalac.mesmer.otelextension.instrumentations.akka.persistence.AkkaPersistenceAgent;
 
 import java.util.Arrays;
 import java.util.List;
@@ -22,11 +22,11 @@ public class MesmerAkkaPersistenceInstrumentationModule extends InstrumentationM
   @Override
   public List<String> getAdditionalHelperClassNames() {
     return Arrays.asList(
-            "io.scalac.mesmer.agent.akka.persistence.impl.JournalInteractionsAdvice$",
-            "io.scalac.mesmer.agent.akka.persistence.impl.StoringSnapshotAdvice$",
-            "io.scalac.mesmer.agent.akka.persistence.impl.RecoveryStartedAdvice$",
-            "io.scalac.mesmer.agent.akka.persistence.impl.PersistingEventSuccessAdvice$",
-            "io.scalac.mesmer.agent.akka.persistence.impl.RecoveryCompletedAdvice$"
+            "io.scalac.mesmer.otelextension.instrumentations.akka.persistence.impl.JournalInteractionsAdvice$",
+            "io.scalac.mesmer.otelextension.instrumentations.akka.persistence.impl.StoringSnapshotAdvice$",
+            "io.scalac.mesmer.otelextension.instrumentations.akka.persistence.impl.RecoveryStartedAdvice$",
+            "io.scalac.mesmer.otelextension.instrumentations.akka.persistence.impl.PersistingEventSuccessAdvice$",
+            "io.scalac.mesmer.otelextension.instrumentations.akka.persistence.impl.RecoveryCompletedAdvice$"
     );
   }
 }
