+++ otel-extension/src/main/scala/io/scalac/mesmer/otelextension/instrumentations/akka/persistence/impl/JournalInteractionsAdvice.scala
@@ -1,23 +1,22 @@
-package io.scalac.mesmer.agent.akka.persistence.impl
+package io.scalac.mesmer.otelextension.instrumentations.akka.persistence.impl
 
 import akka.actor.typed.scaladsl.ActorContext
 import akka.persistence.PersistentRepr
 import net.bytebuddy.asm.Advice._
 
 import io.scalac.mesmer.core.event.EventBus
-import io.scalac.mesmer.core.event.PersistenceEvent.PersistingEventFinished
+import io.scalac.mesmer.core.event.PersistenceEvent.PersistingEventStarted
 import io.scalac.mesmer.core.model._
 import io.scalac.mesmer.core.util.Timestamp
 
-object PersistingEventSuccessAdvice {
+object JournalInteractionsAdvice {
 
   @OnMethodEnter
-  def onWriteSuccess(@Argument(0) context: ActorContext[_], @Argument(1) event: PersistentRepr): Unit = {
-    val path: Path = context.self.path.toPath
-
+  def onWriteInitiated(@Argument(0) context: ActorContext[_], @Argument(2) event: PersistentRepr): Unit = {
+    val path = context.self.path.toPath
     EventBus(context.system)
       .publishEvent(
-        PersistingEventFinished(
+        PersistingEventStarted(
           path,
           event.persistenceId,
           event.sequenceNr,
@@ -25,4 +24,5 @@ object PersistingEventSuccessAdvice {
         )
       )
   }
+
 }
