+++ otel-extension/src/main/scala/io/scalac/mesmer/otelextension/instrumentations/akka/persistence/impl/RecoveryCompletedAdvice.scala
@@ -1,10 +1,12 @@
-package io.scalac.mesmer.agent.akka.persistence.impl
+package io.scalac.mesmer.otelextension.instrumentations.akka.persistence.impl
 
 import java.lang.invoke.MethodHandle
 
-import _root_.akka.actor.typed.scaladsl.ActorContext
-import _root_.akka.persistence.typed.PersistenceId
-import net.bytebuddy.asm.Advice._
+import akka.actor.typed.scaladsl.ActorContext
+import akka.persistence.typed.PersistenceId
+import net.bytebuddy.asm.Advice.Argument
+import net.bytebuddy.asm.Advice.OnMethodEnter
+import net.bytebuddy.asm.Advice.This
 
 import io.scalac.mesmer.core.event.EventBus
 import io.scalac.mesmer.core.event.PersistenceEvent.RecoveryFinished
